Refactor CreateReadwriteSplittingRuleStatementUpdaterTest
